recents rearrangement recents + fixes
Recents was and is a crap...already from AOSP...whatever
to avoid animation conflicts with the RAM bar it is moved up.
Due a lot user requests the clear recents is not bottom left
Recents container is now reduced and on bottom off ram bar
so that no visual overlap glitch exist any more. As well the
recents scrollview is more stable and now small 1dp glitches like
the original AOSP implementation. Code cleanup
and some code arrangements.

cheers

Change-Id: I3d4d880a6fa41575b2c9a4de06d289551aa5436d