Check and fail early if requested wallpaper size exceeds maximum texture size
This fixes an issue where OpenGL initialization succeeds but buffer
allocation fails because the requested wallpaper size
is too large (or otherwise unsupported) by the graphics hardware. This
fixes an issue where SystemUI crashes constantly
on the PandaBoard when connected to a full HD display. Tested only on
PandaBoard, no access to alternative hardware.
Signed-off-by: Wim Vander Schelden <wim.vander.schelden@philips.com>

Change-Id: I8d2e1ae9fd9772977c4e365f23f2f58bbca3787c